gitignore

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double height, int weight) {
+    public long calculate(double height, int weight) {
         long calculate = (long) (weight / (height * height));
         return calculate;
     }
